--- model/contacts.json
@@ -1,1 +1,0 @@
-[{"id":3,"name":"John","email":"john@john.com","phone":"222-22-22"},{"id":4,"name":"Wylie Pope","email":"est@utquamvel.net","phone":"(692) 802-2949"},{"id":5,"name":"Cyrus Jackson","email":"nibh@semsempererat.com","phone":"(501) 472-5218"},{"id":6,"name":"Yurii","email":"Yurii@mail.com","phone":"222 333 2"},{"id":7,"name":"Reuben Henry","email":"pharetra.ut@dictum.co.uk","phone":"(715) 598-5792"},{"id":8,"name":"poly","email":"poly@mail.net","phone":"111-222-333-555-666-777"},{"id":9,"name":"Thomas Lucas","email":"nec@Nulla.com","phone":"(704) 398-7993"},{"id":10,"name":"aaa","email":"aaa@mail.com","phone":"222 333 444"},{"id":11,"name":"John","email":"john@john.com","phone":"222-22-22"},{"id":12,"name":"John","email":"john@john.com","phone":"222-22-22"}]
